+++ src/styles/themes/accent.js
@@ -1,21 +1,23 @@
 import createMuiTheme from '@material-ui/core/styles/createMuiTheme'
 import memoize from 'lodash/memoize'
-import merge from 'deepmerge'
+import { deepmerge } from '@material-ui/utils'
+
 import accent from '../palettes/accent'
-import { typography, errorStateColours } from './common'
+import { typography, errorStateColours, zIndex } from './common'
 
 /**
  * This theme contains 'accent' coloured swatches
  */
 const theme = createMuiTheme(
-  merge(
+  deepmerge(
     {
       palette: {
         // mui colours
         primary: { main: accent[700] },
         secondary: { main: accent[600] }
       },
-      typography
+      typography,
+      zIndex
     }, errorStateColours
   )
 )