+++ src/styles/themes/core.js
@@ -1,22 +1,26 @@
 import createMuiTheme from '@material-ui/core/styles/createMuiTheme'
 import memoize from 'lodash/memoize'
-import merge from 'deepmerge'
-import { typography, errorStateColours } from './common'
+import { deepmerge } from '@material-ui/utils'
+
 import core from '../palettes/core'
+import { typography, errorStateColours, zIndex } from './common'
+
 /**
  *  This theme contains 'core' coloured swatches
  */
 
 const theme = createMuiTheme(
-  merge(
+  deepmerge(
     {
       palette: {
       // mui colours
         primary: { main: core[600] },
         secondary: { main: core[400] }
       },
-      typography
-    }, errorStateColours)
+      typography,
+      zIndex
+    }, errorStateColours
+  )
 )
 
 export default memoize(() => theme)